+++ spring/src/test/java/com/influxdb/spring/influx/InfluxDB2AutoConfigurationTest.java
@@ -23,10 +23,10 @@ package com.influxdb.spring.influx;
 
 import java.util.List;
 import java.util.concurrent.TimeUnit;
-
 import javax.annotation.Nonnull;
 
 import com.influxdb.client.InfluxDBClient;
+import com.influxdb.client.reactive.InfluxDBClientReactive;
 
 import okhttp3.OkHttpClient;
 import okhttp3.Protocol;
@@ -35,6 +35,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.platform.runner.JUnitPlatform;
 import org.junit.runner.RunWith;
 import org.springframework.boot.autoconfigure.AutoConfigurations;
+import org.springframework.boot.test.context.FilteredClassLoader;
 import org.springframework.boot.test.context.assertj.AssertableApplicationContext;
 import org.springframework.boot.test.context.runner.ApplicationContextRunner;
 import org.springframework.context.annotation.Bean;
@@ -51,6 +52,7 @@ import retrofit2.Retrofit;
 class InfluxDB2AutoConfigurationTest {
 
     private final ApplicationContextRunner contextRunner = new ApplicationContextRunner()
+            .withClassLoader(new FilteredClassLoader(InfluxDBClientReactive.class))
             .withConfiguration(AutoConfigurations.of(InfluxDB2AutoConfiguration.class));
 
     @Test
@@ -62,15 +64,15 @@ class InfluxDB2AutoConfigurationTest {
     @Test
     public void influxDBClientCanBeCustomized() {
         this.contextRunner
-                .withPropertyValues("spring.influx2.url=http://localhost:8086/",
-                        "spring.influx2.password:password", "spring.influx2.username:username")
+                .withPropertyValues("influx.url=http://localhost:8086/",
+                        "influx.password:password", "influx.username:username")
                 .run(((context) -> Assertions.assertThat(context.getBeansOfType(InfluxDBClient.class))
                         .hasSize(1)));
     }
 
     @Test
     public void influxDBClientCanBeCreatedWithoutCredentials() {
-        this.contextRunner.withPropertyValues("spring.influx2.url=http://localhost:8086/")
+        this.contextRunner.withPropertyValues("influx.url=http://localhost:8086/")
                 .run((context) -> {
                     Assertions.assertThat(context.getBeansOfType(InfluxDBClient.class)).hasSize(1);
                     int readTimeout = getReadTimeoutProperty(context);
@@ -82,7 +84,7 @@ class InfluxDB2AutoConfigurationTest {
     public void influxDBClientWithOkHttpClientBuilderProvider() {
         this.contextRunner
                 .withUserConfiguration(CustomOkHttpClientBuilderProviderConfig.class)
-                .withPropertyValues("spring.influx2.url=http://localhost:8086/", "spring.influx2.token:token")
+                .withPropertyValues("influx.url=http://localhost:8086/", "influx.token:token")
                 .run((context) -> {
                     Assertions.assertThat(context.getBeansOfType(InfluxDBClient.class)).hasSize(1);
                     int readTimeout = getReadTimeoutProperty(context);
@@ -92,7 +94,7 @@ class InfluxDB2AutoConfigurationTest {
 
     @Test
     public void influxDBClientWithReadTimeout() {
-        this.contextRunner.withPropertyValues("spring.influx2.url=http://localhost:8086/", "spring.influx2.readTimeout=13s")
+        this.contextRunner.withPropertyValues("influx.url=http://localhost:8086/", "influx.readTimeout=13s")
                 .run((context) -> {
                     Assertions.assertThat(context.getBeansOfType(InfluxDBClient.class)).hasSize(1);
                     int readTimeout = getReadTimeoutProperty(context);
@@ -101,8 +103,23 @@ class InfluxDB2AutoConfigurationTest {
     }
 
     @Test
+    public void influxDBClientReactive() {
+        ApplicationContextRunner contextRunner = new ApplicationContextRunner()
+                .withPropertyValues("influx.url=http://localhost:8086/")
+                .withConfiguration(AutoConfigurations.of(InfluxDB2AutoConfiguration.class));
+
+        contextRunner
+                .run(((context) -> Assertions.assertThat(context.getBeansOfType(InfluxDBClientReactive.class))
+                .hasSize(1)));
+
+        contextRunner
+                .run(((context) -> Assertions.assertThat(context.getBeansOfType(InfluxDBClient.class))
+                .hasSize(0)));
+    }
+
+    @Test
     public void protocolVersion() {
-        this.contextRunner.withPropertyValues("spring.influx2.url=http://localhost:8086/", "spring.influx2.token:token")
+        this.contextRunner.withPropertyValues("influx.url=http://localhost:8086/", "spring.influx2.token:token")
                 .run((context) -> {
                     List<Protocol> protocols = getOkHttpClient(context).protocols();
                     Assertions.assertThat(protocols).hasSize(1);